- Fixed bug in Distance2D   * distance between line segment and point   * bad unit test hid the bug - Removed excessive typecast from FitLine

+++ src/georegression/fitting/line/FitLine_F32.java
@@ -67,7 +67,7 @@ public class FitLine_F32 {
 			bottom += dy*dy - dx*dx;
 		}
 
-		ret.angle = (float) (float)Math.atan2(-2.0f*top , bottom)/2.0f;
+		ret.angle = (float)Math.atan2(-2.0f*top , bottom)/2.0f;
 		ret.distance = (float)( meanX*Math.cos(ret.angle) + meanY*Math.sin(ret.angle));
 
 		return ret;
